Odin Data Amendments
Tracking Table Fix 1: Methodology change. Compare between the max number of comments first date and compare it with the last date. If it's more than 3 days, denote the submission as close.
Tracking Table Fix 2: For some of the earlier submissions that have ended early on and did not continue updating, we closed the submissions